chore (teams): refactor team page
