+++ src/app/home/my-team/page.tsx
@@ -2,39 +2,16 @@
 import { Layout, Button, CustomLink, EmptyCard } from "@/common";
 import { AddIcon, UpLoadIcon } from "@/common/Icons";
 import { useStore } from "@/models/root.store";
+import DataTeam from "./DataTeam";
+import EmptyTeam from "./EmptyTeam";
 
 export default function MyTeam() {
   const {
-    aside: { setAside },
+    members: { members },
   } = useStore();
   return (
-    <Layout className="border-2 shadow-sm border-border rounded-md grid place-items-center w-[98%] ">
-      <EmptyCard
-        imageBottom="/girl.svg"
-        paragraph="You havnet't loaded any  employees yet."
-        altImage="Girl"
-      >
-        <div className="flex gap-2 ">
-          <Button
-            body="Load Team Members"
-            icon={<AddIcon />}
-            onClick={() => {
-              setAside("LoadStock");
-            }}
-            variant="secondary"
-            size="big"
-            className="rounded-md"
-          />
-          <CustomLink
-            variant="primary"
-            size="big"
-            className="rounded-md flex gap-2"
-            href="#"
-          >
-            <UpLoadIcon /> Add Team Memeber
-          </CustomLink>
-        </div>
-      </EmptyCard>
-    </Layout>
+    <div>
+      {members.length ? <DataTeam /> : <EmptyTeam />}
+    </div>
   );
 }
